Move Health system to Core.
