+++ engine/src/main/java/org/terasology/logic/characters/CharacterSystem.java
@@ -22,8 +22,8 @@ import org.slf4j.LoggerFactory;
 import org.terasology.engine.Time;
 import org.terasology.entitySystem.entity.EntityManager;
 import org.terasology.entitySystem.entity.EntityRef;
+import org.terasology.entitySystem.event.EventPriority;
 import org.terasology.entitySystem.event.ReceiveEvent;
-import org.terasology.entitySystem.prefab.Prefab;
 import org.terasology.entitySystem.systems.BaseComponentSystem;
 import org.terasology.entitySystem.systems.RegisterMode;
 import org.terasology.entitySystem.systems.RegisterSystem;
@@ -31,13 +31,14 @@ import org.terasology.entitySystem.systems.UpdateSubscriberSystem;
 import org.terasology.input.binds.interaction.AttackButton;
 import org.terasology.logic.characters.events.ActivationRequest;
 import org.terasology.logic.characters.events.ActivationRequestDenied;
+import org.terasology.logic.characters.events.AttackEvent;
 import org.terasology.logic.characters.events.AttackRequest;
 import org.terasology.logic.characters.events.DeathEvent;
 import org.terasology.logic.characters.interactions.InteractionUtil;
 import org.terasology.logic.common.ActivateEvent;
 import org.terasology.logic.common.DisplayNameComponent;
 import org.terasology.logic.health.DestroyEvent;
-import org.terasology.logic.health.DoDamageEvent;
+import org.terasology.logic.health.DoDestroyEvent;
 import org.terasology.logic.health.EngineDamageTypes;
 import org.terasology.logic.inventory.ItemComponent;
 import org.terasology.logic.location.LocationComponent;
@@ -49,6 +50,8 @@ import org.terasology.physics.HitResult;
 import org.terasology.physics.Physics;
 import org.terasology.physics.StandardCollisionGroup;
 import org.terasology.registry.In;
+import org.terasology.world.block.BlockComponent;
+import org.terasology.world.block.regions.ActAsBlockComponent;
 
 /**
  */
@@ -70,13 +73,12 @@ public class CharacterSystem extends BaseComponentSystem implements UpdateSubscr
     private Time time;
 
     @ReceiveEvent(components = {CharacterComponent.class})
-    public void onDeath(DestroyEvent event, EntityRef entity) {
+    public void onDeath(DoDestroyEvent event, EntityRef entity) {
         CharacterComponent character = entity.getComponent(CharacterComponent.class);
         character.controller.send(new DeathEvent());
         // TODO: Don't just destroy, ragdoll or create particle effect or something (possible allow another system to handle)
         //entity.removeComponent(CharacterComponent.class);
         //entity.removeComponent(CharacterMovementComponent.class);
-        entity.destroy();
     }
 
 
@@ -105,8 +107,10 @@ public class CharacterSystem extends BaseComponentSystem implements UpdateSubscr
         event.consume();
     }
 
-    @ReceiveEvent(components = {CharacterComponent.class, LocationComponent.class})
+
+    @ReceiveEvent(components = LocationComponent.class, netFilter = RegisterMode.AUTHORITY)
     public void onAttackRequest(AttackRequest event, EntityRef character) {
+        // if an item is used,  make sure this entity is allowed to attack with it
         if (event.getItem().exists()) {
             if (!character.equals(event.getItem().getOwner())) {
                 return;
@@ -122,20 +126,20 @@ public class CharacterSystem extends BaseComponentSystem implements UpdateSubscr
         HitResult result = physics.rayTrace(originPos, direction, characterComponent.interactionRange, Sets.newHashSet(character), DEFAULTPHYSICSFILTER);
 
         if (result.isHit()) {
-            int damage = 1;
-            Prefab damageType = EngineDamageTypes.PHYSICAL.get();
-            // Calculate damage from item
-            ItemComponent item = event.getItem().getComponent(ItemComponent.class);
-            if (item != null) {
-                damage = item.baseDamage;
-                if (item.damageType != null) {
-                    damageType = item.damageType;
-                }
-            }
-
-            result.getEntity().send(new DoDamageEvent(damage, damageType, character, event.getItem()));
+            result.getEntity().send(new AttackEvent(character, event.getItem()));
         }
     }
+
+    @ReceiveEvent(priority = EventPriority.PRIORITY_TRIVIAL, netFilter = RegisterMode.AUTHORITY)
+    public void onAttackBlock(AttackEvent event, EntityRef entityRef, BlockComponent blockComponent) {
+        entityRef.send(new DestroyEvent(event.getInstigator(), event.getDirectCause(), EngineDamageTypes.PHYSICAL.get()));
+    }
+
+    @ReceiveEvent(priority = EventPriority.PRIORITY_TRIVIAL, netFilter = RegisterMode.AUTHORITY)
+    public void onAttackBlock(AttackEvent event, EntityRef entityRef, ActAsBlockComponent actAsBlockComponent) {
+        entityRef.send(new DestroyEvent(event.getInstigator(), event.getDirectCause(), EngineDamageTypes.PHYSICAL.get()));
+    }
+
 
     @ReceiveEvent(components = {CharacterComponent.class, LocationComponent.class}, netFilter = RegisterMode.AUTHORITY)
     public void onActivationRequest(ActivationRequest event, EntityRef character) {
