+++ engine/src/main/java/org/terasology/logic/players/PlayerSystem.java
@@ -30,9 +30,6 @@ import org.terasology.logic.characters.CharacterComponent;
 import org.terasology.logic.console.commandSystem.annotations.Command;
 import org.terasology.logic.console.commandSystem.annotations.CommandParam;
 import org.terasology.logic.console.commandSystem.annotations.Sender;
-import org.terasology.logic.health.DestroyEvent;
-import org.terasology.logic.health.EngineDamageTypes;
-import org.terasology.logic.health.HealthComponent;
 import org.terasology.logic.location.Location;
 import org.terasology.logic.location.LocationComponent;
 import org.terasology.logic.permission.PermissionManager;
@@ -203,16 +200,6 @@ public class PlayerSystem extends BaseComponentSystem implements UpdateSubscribe
             updateRelevanceEntity(entity, ViewDistance.LEGALLY_BLIND.getChunkDistance());
             SpawningClientInfo info = new SpawningClientInfo(entity, spawnPosition);
             clientsPreparingToSpawn.add(info);
-        }
-    }
-
-    @Command(value = "kill", shortDescription = "Reduce the player's health to zero", runOnServer = true,
-            requiredPermission = PermissionManager.NO_PERMISSION)
-    public void killCommand(@Sender EntityRef client) {
-        ClientComponent clientComp = client.getComponent(ClientComponent.class);
-        HealthComponent health = clientComp.character.getComponent(HealthComponent.class);
-        if (health != null) {
-            clientComp.character.send(new DestroyEvent(clientComp.character, EntityRef.NULL, EngineDamageTypes.DIRECT.get()));
         }
     }
