+++ engine/src/main/java/org/terasology/world/block/entity/BlockEntitySystem.java
@@ -18,23 +18,16 @@ package org.terasology.world.block.entity;
 import org.terasology.audio.AudioManager;
 import org.terasology.audio.StaticSound;
 import org.terasology.audio.events.PlaySoundEvent;
-import org.terasology.entitySystem.entity.EntityBuilder;
 import org.terasology.entitySystem.entity.EntityManager;
 import org.terasology.entitySystem.entity.EntityRef;
 import org.terasology.entitySystem.event.EventPriority;
 import org.terasology.entitySystem.event.ReceiveEvent;
 import org.terasology.entitySystem.systems.BaseComponentSystem;
-import org.terasology.entitySystem.systems.RegisterMode;
 import org.terasology.entitySystem.systems.RegisterSystem;
-import org.terasology.logic.health.BeforeDamagedEvent;
 import org.terasology.logic.health.DoDestroyEvent;
-import org.terasology.logic.health.FullHealthEvent;
-import org.terasology.logic.health.OnDamagedEvent;
 import org.terasology.logic.inventory.events.DropItemEvent;
 import org.terasology.logic.inventory.events.GiveItemEvent;
 import org.terasology.logic.location.LocationComponent;
-import org.terasology.logic.particles.BlockParticleEffectComponent;
-import org.terasology.math.geom.Vector3f;
 import org.terasology.math.geom.Vector3i;
 import org.terasology.physics.events.ImpulseEvent;
 import org.terasology.registry.In;
@@ -45,8 +38,6 @@ import org.terasology.world.block.Block;
 import org.terasology.world.block.BlockComponent;
 import org.terasology.world.block.BlockManager;
 import org.terasology.world.block.entity.damage.BlockDamageModifierComponent;
-import org.terasology.world.block.entity.damage.BlockDamagedComponent;
-import org.terasology.world.block.family.BlockFamily;
 import org.terasology.world.block.items.BlockItemFactory;
 import org.terasology.world.block.items.OnBlockToItem;
 import org.terasology.world.block.regions.ActAsBlockComponent;
@@ -177,94 +168,4 @@ public class BlockEntitySystem extends BaseComponentSystem {
         item.send(new ImpulseEvent(random.nextVector3f(30.0f)));
     }
 
-    @ReceiveEvent
-    public void beforeDamaged(BeforeDamagedEvent event, EntityRef blockEntity, BlockComponent blockComp) {
-        if (!blockComp.getBlock().isDestructible()) {
-            event.consume();
-        }
-    }
-
-    @ReceiveEvent
-    public void beforeDamaged(BeforeDamagedEvent event, EntityRef blockEntity, ActAsBlockComponent blockComp) {
-        if (blockComp.block != null && !blockComp.block.getArchetypeBlock().isDestructible()) {
-            event.consume();
-        }
-    }
-
-    @ReceiveEvent(components = {BlockDamagedComponent.class})
-    public void onRepaired(FullHealthEvent event, EntityRef entity) {
-        entity.removeComponent(BlockDamagedComponent.class);
-    }
-
-    @ReceiveEvent
-    public void onDamaged(OnDamagedEvent event, EntityRef entity, BlockComponent blockComponent, LocationComponent locComp) {
-        onDamagedCommon(event, entity, blockComponent.getBlock().getBlockFamily(), locComp.getWorldPosition());
-        if (!entity.hasComponent(BlockDamagedComponent.class)) {
-            entity.addComponent(new BlockDamagedComponent());
-        }
-    }
-
-    @ReceiveEvent
-    public void onDamaged(OnDamagedEvent event, EntityRef entity, ActAsBlockComponent blockComponent, LocationComponent locComp) {
-        if (blockComponent.block != null) {
-            onDamagedCommon(event, entity, blockComponent.block, locComp.getWorldPosition());
-        }
-
-    }
-
-    public void onDamagedCommon(OnDamagedEvent event, EntityRef entity, BlockFamily blockFamily, Vector3f location) {
-        BlockDamageModifierComponent blockDamageSettings = event.getType().getComponent(BlockDamageModifierComponent.class);
-        boolean skipDamageEffects = false;
-        if (blockDamageSettings != null) {
-            skipDamageEffects = blockDamageSettings.skipPerBlockEffects;
-        }
-        if (!skipDamageEffects) {
-            onPlayBlockDamageCommon(blockFamily, location);
-        }
-    }
-
-    private void onPlayBlockDamageCommon(BlockFamily family, Vector3f location) {
-        EntityBuilder builder = entityManager.newBuilder("engine:defaultBlockParticles");
-        builder.getComponent(LocationComponent.class).setWorldPosition(location);
-        builder.getComponent(BlockParticleEffectComponent.class).blockType = family;
-        builder.build();
-
-        if (family.getArchetypeBlock().isDebrisOnDestroy()) {
-            EntityBuilder dustBuilder = entityManager.newBuilder("engine:dustEffect");
-            dustBuilder.getComponent(LocationComponent.class).setWorldPosition(location);
-            dustBuilder.build();
-        }
-
-        BlockSounds sounds = family.getArchetypeBlock().getSounds();
-        if (!sounds.getDigSounds().isEmpty()) {
-            StaticSound sound = random.nextItem(sounds.getDigSounds());
-            audioManager.playSound(sound, location);
-        }
-    }
-
-    @ReceiveEvent(netFilter = RegisterMode.AUTHORITY)
-    public void beforeDamage(BeforeDamagedEvent event, EntityRef entity, BlockComponent blockComp) {
-        beforeDamageCommon(event, blockComp.getBlock());
-    }
-
-    @ReceiveEvent(netFilter = RegisterMode.AUTHORITY)
-    public void beforeDamage(BeforeDamagedEvent event, EntityRef entity, ActAsBlockComponent blockComp) {
-        if (blockComp.block != null) {
-            beforeDamageCommon(event, blockComp.block.getArchetypeBlock());
-        }
-    }
-
-    private void beforeDamageCommon(BeforeDamagedEvent event, Block block) {
-        if (event.getDamageType() != null) {
-            BlockDamageModifierComponent blockDamage = event.getDamageType().getComponent(BlockDamageModifierComponent.class);
-            if (blockDamage != null) {
-                BlockFamily blockFamily = block.getBlockFamily();
-                for (String category : blockFamily.getCategories()) {
-                    if (blockDamage.materialDamageMultiplier.containsKey(category)) {
-                        event.multiply(blockDamage.materialDamageMultiplier.get(category));
-                    }
-                }
-            }
-        }
-    }
 }
