+++ engine/src/main/resources/assets/prefabs/player/player.prefab
@@ -7,15 +7,6 @@
     "Network" : {
         "replicateMode" : "ALWAYS"
     },
-    "Health" : {
-        "currentHealth" : 100,
-        "maxHealth" : 100,
-        "regenRate" : 3,
-        "waitBeforeRegen" : 10,
-        "excessSpeedDamageMultiplier" : 5,
-        "fallingDamageSpeedThreshold" : 18,
-        "horizontalDamageSpeedThreshold" : 50
-    },
     "CharacterMovement" : {
         "groundFriction" : 16,
         "maxGroundSpeed" : 3,
