+++ engine/src/main/resources/assets/ui/hud/toolbar.ui
@@ -6,24 +6,6 @@
         "contents" : [
             {
                 "type" : "UIIconBar",
-                "id" : "healthBar",
-                "family" : "healthBar",
-                "icon" : "engine:icons#redHeart",
-                "halfIconMode" : "split",
-                "spacing" : 2,
-                "maxIcons" : 10,
-                "layoutInfo" : {
-                    "use-content-width" : true,
-                    "use-content-height" : true,
-                    "position-horizontal-center": {},
-                    "position-bottom" : {
-                        "target": "BOTTOM",
-                        "offset": 60
-                    }
-                }
-            },
-            {
-                "type" : "UIIconBar",
                 "id" : "breathBar",
                 "icon" : "engine:icons#bubble",
                 "family" : "breathBar",
@@ -52,8 +34,6 @@
                 "type" : "UICrosshair",
                 "id" : "crosshair",
                 "crosshairIcon" : "engine:gui#crosshair",
-                "chargeStages" : ["engine:gui#crosshairCharge1","engine:gui#crosshairCharge2","engine:gui#crosshairCharge3","engine:gui#crosshairCharge4",
-                                  "engine:gui#crosshairCharge5","engine:gui#crosshairCharge6","engine:gui#crosshairCharge7","engine:gui#crosshairCharge8"],
                 "layoutInfo" : {
                     "use-content-width" : true,
                     "use-content-height" : true,
