+++ modules/Core/assets/prefabs/doorRegion.prefab
@@ -1,0 +1,26 @@
+{
+  "DisplayName": {
+    "name": "Door"
+  },
+  "Door": {
+    "topBlockFamily": "DoorTop",
+    "bottomBlockFamily": "DoorBottom",
+    "openSound": "DoorOpen",
+    "closeSound": "DoorClose"
+  },
+  "Health": {
+    "currentHealth": 5,
+    "maxHealth": 5,
+    "regenRate": 1,
+    "waitBeforeRegen": 2,
+    "destroyEntityOnNoHealth": true
+  },
+  "BlockDropGrammar": {
+    "itemDrops": [
+      "Core:door"
+    ]
+  },
+  "ActAsBlock": {
+    "block": "core:DoorTop"
+  }
+}
