+++ modules/Core/deltas/engine/prefabs/player/player.prefab
@@ -44,5 +44,15 @@
     ]
   },
   "SelectedInventorySlot": {
+  },
+  "Health": {
+    "currentHealth": 100,
+    "maxHealth": 100,
+    "regenRate": 3,
+    "waitBeforeRegen": 10,
+    "excessSpeedDamageMultiplier": 5,
+    "fallingDamageSpeedThreshold": 18,
+    "horizontalDamageSpeedThreshold": 50,
+    "destroyEntityOnNoHealth": true
   }
 }
