+++ modules/Core/src/main/java/org/terasology/logic/health/BeforeDamagedEvent.java
