+++ modules/Core/src/main/java/org/terasology/logic/health/BeforeHealEvent.java
