+++ modules/Core/src/main/java/org/terasology/logic/health/BlockDamagedComponent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 MovingBlocks
+ * Copyright 2016 MovingBlocks
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,12 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.world.block.entity.damage;
+package org.terasology.logic.health;
 
 import org.terasology.entitySystem.Component;
 import org.terasology.world.block.ForceBlockActive;
 
 /**
+ * This component is added by the authority to temporary block entities so that they can persist for a bit while the block is being damaged.
  */
 @ForceBlockActive
 public class BlockDamagedComponent implements Component {
